reorganized in preparation for multiple layouts. everything is now in folders based on their file type, and files associated with a specific room are in their own folder.
