+++ lab_021/index.html
@@ -1,0 +1,94 @@
+<html>
+<head>
+<link rel="stylesheet" type="text/css" href="/index.css">
+<script src="/socket.io/socket.io.js"></script>
+<script src="/signin.js"></script>
+<title>SDMS Sign In: Room 021</title>
+</head>
+<body>
+
+
+<ul class="tabs">
+    <li>
+        <input type="radio" checked name="tabs" id="signin" onClick="switchTabStyles(this.id)">
+            <label for="signin">Sign In</label>
+            <div class="tab-content">
+                1. Type your Student ID or scan your barcode.<br>
+                <input id="sid" type="text" pattern="[0-9]*" maxlength="8" autofocus /><br>
+                2. Choose a computer ---><br>
+                3. Click Go:
+                <div class="button" onclick="signin()">Go</div> 
+            </div>
+        </li>
+    <li>
+        <input type="radio" name="tabs" id="signout" onClick="switchTabStyles(this.id)">
+            <label for="signout">Sign Out</label>
+            <div class="tab-content">
+                1. Click on your name ---><br>
+                2. Select your destination:<br>
+                <div id="destination-container">
+                <div class="row-destinations1">
+                	<div class="destination" id="cafeteria" onClick="chooseDestination(this.id)">Cafeteria</div>
+                	<div class="destination" id="classroom" onClick="chooseDestination(this.id)">Classroom</div>
+                	<div class="destination" id="mediacenter" onClick="chooseDestination(this.id)">Media Center</div>
+                	<div class="destination" id="office" onClick="chooseDestination(this.id)">Office</div>
+                </div>
+                <div class="row-destinations2">
+	                <div class="destination" id="nurse" onClick="chooseDestination(this.id)">Nurse</div>
+	                <div class="destination" id="locker" onClick="chooseDestination(this.id)">Locker</div>
+	                <div class="destination" id="bathroom" onClick="chooseDestination(this.id)">Bathroom</div>
+	                <div class="destination" id="water" onClick="chooseDestination(this.id)">Water</div>
+	            </div>
+	            </div>
+	            3. Click Go: 
+                <div class="button" onclick="signout()">Go</div>
+            </div>
+    </li>
+</ul>
+<div id="map" class="signin">
+ <div id="row1">
+     <div id="1" class="computer" onClick="chooseComputer(this.id)">1</div>
+     <div id="2" class="computer" onClick="chooseComputer(this.id)">2</div>
+     <div id="3" class="computer" onClick="chooseComputer(this.id)">3</div>
+     <div id="4" class="computer" onClick="chooseComputer(this.id)">4</div>
+     <div id="5" class="computer" onClick="chooseComputer(this.id)">5</div>
+     <div id="6" class="computer" onClick="chooseComputer(this.id)">6</div>
+     <div id="7" class="computer" onClick="chooseComputer(this.id)">7</div>
+     <div id="8" class="computer" onClick="chooseComputer(this.id)">8</div>
+ </div>
+ <div id="row2">
+     <div id="9" class="computer" onClick="chooseComputer(this.id)">9</div>
+     <div id="10" class="computer" onClick="chooseComputer(this.id)">10</div>
+     <div id="11" class="computer" onClick="chooseComputer(this.id)">11</div>
+     <div id="12" class="computer" onClick="chooseComputer(this.id)">12</div>
+     <div id="13" class="computer" onClick="chooseComputer(this.id)">13</div>
+     <div id="14" class="computer" onClick="chooseComputer(this.id)">14</div>
+     <div id="15" class="computer" onClick="chooseComputer(this.id)">15</div>
+     <div id="16" class="computer" onClick="chooseComputer(this.id)">16</div>
+ </div>
+ <div id="row3">
+     <div id="17" class="computer" onClick="chooseComputer(this.id)">17</div>
+     <div id="18" class="computer" onClick="chooseComputer(this.id)">18</div>
+     <div id="19" class="computer" onClick="chooseComputer(this.id)">19</div>
+     <div id="20" class="computer" onClick="chooseComputer(this.id)">20</div>
+     <div id="21" class="computer" onClick="chooseComputer(this.id)">21</div>
+     <div id="22" class="computer" onClick="chooseComputer(this.id)">22</div>
+     <div id="23" class="computer" onClick="chooseComputer(this.id)">23</div>
+     <div id="24" class="computer" onClick="chooseComputer(this.id)">24</div>
+ </div>
+ <div id="row4">
+     <div id="25" class="computer" onClick="chooseComputer(this.id)">25</div>
+     <div id="26" class="computer" onClick="chooseComputer(this.id)">26</div>
+     <div id="27" class="computer" onClick="chooseComputer(this.id)">27</div>
+     <div id="28" class="computer" onClick="chooseComputer(this.id)">28</div>
+     <div id="29" class="computer" onClick="chooseComputer(this.id)">29</div>
+     <div id="30" class="computer" onClick="chooseComputer(this.id)">30</div>
+     <div id="31" class="computer" onClick="chooseComputer(this.id)">31</div>
+     <div id="0" class="computer taken">PC Computer</div>
+ </div>
+</div>
+
+
+</body>
+</html>
+
